% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASSO_wrapper.R
\name{fitLASSO_prox_Nesterov}
\alias{fitLASSO_prox_Nesterov}
\title{Title}
\usage{
fitLASSO_prox_Nesterov(X, Y, lambda, beta_start = NULL, eps = 1e-04, s = 0.01)
}
\arguments{
\item{X}{}

\item{lambda}{}

\item{beta_start}{}

\item{eps}{}

\item{s}{}
}
\description{
Title
}
