% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_wrapper.R
\name{LRMultiClass}
\alias{LRMultiClass}
\title{Title}
\usage{
LRMultiClass(X, y, beta_init = NULL, numIter = 50, eta = 0.1, lambda = 1)
}
\arguments{
\item{X}{}

\item{y}{}

\item{beta_init}{}

\item{numIter}{}

\item{eta}{}

\item{lambda}{}
}
\description{
Title
}
\examples{
# Give example
}
